+++ src/components/Startsida/index.ts
@@ -1,0 +1,2 @@
+export * from './porttelefon';
+export * from './kommandeAktiviteter';
